0529夕方時点の進捗です！ソート機能の一部まで使えるようになりました。集計系はなぜか0.0が数値に入るので原因調査中です。

+++ src/main/java/com/example/demo/MyService.java
@@ -39,7 +39,7 @@ public class MyService {
 	}
 
 	//	収支毎に日付でソート
-	public List<Payments> getKindSortedByDate(LocalDate date, int user) {
+	public List<Payments> getKindSortedByDate(LocalDate date, long user) {
 		List<Payments> Payments;
 		if (user != 0) {
 			Payments = paymentsRepository.findByDateAndUserId(date, user);
@@ -52,7 +52,7 @@ public class MyService {
 	}
 
 	//	収支毎に金額でソート
-	public List<Payments> getKindSortedByAmount(int amount, int user) {
+	public List<Payments> getKindSortedByAmount(int amount, long user) {
 		List<Payments> Payments;
 		if (user != 0) {
 			Payments = paymentsRepository.findByAmountAndUserId(amount, user);
